+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Change log
 
 - Fixed a crash where apps without ``required_fields`` could not be
   saved.
+- Added a template snippet based renderer for plugins.
 
 
 `0.4`_ (2016-07-04)